+++ include/Midia.hpp
@@ -39,7 +39,7 @@ class Midia {
 
         //Setters
         void setGenero(Midia::Genero*);
-        void setDuracao(int n);
+        void setDuracao(float n);
         void setAnoLancamento(int n);
 
         //Operators
